1.5-night.rev10
On submit, if it gets "duplicate id error", or any error really, it will
check if the same id and version is prexisting, if it is, then it will
download that.

Stupid JWT thing gets credential errors randomly it seems, so i just
have to wait and try again. I thought it was happening because I was
creating multiple tokens in the same second, but no that's not the case.
Weirdness. So i just keep retyring till it goes through with a 1.5sec
wait.